requirements.txt
streamlit
speedtest-cli
requests
matplotlib
pandas
